Added some safety nets. The PortAudio backend no longer fails as catastrophically. The PortAudio backend now doesn't call Pa_Initialize() after being checked for instantiability. The emitter code has been moved to a subdirectory (emitter). Restructured the audio backend initialization code. Fixed an admittedly unreproducable reported compilation error in MidiFIO::readerOpen().

+++ midifio/readeropen.cpp
@@ -152,7 +152,7 @@ namespace sb {
                         return false;
                 }
             }
-            tracks.push_back(river.tellg()-4L);
+            tracks.push_back(static_cast<size_t>(river.tellg()-4L));
 
             //Index all the tracks.
             if (filetype != 0) {
@@ -181,8 +181,7 @@ namespace sb {
                     }
                     //TODO: For future versions, have a way of checking for an end-of-track event.
                     if(!river.eof())
-                        tracks.push_back(river.tellg()-4L);
-
+                        tracks.push_back(static_cast<size_t>(river.tellg()-4L));
                 }
             }
